fix hotkeys issues
